+++ CHANGELOG.md
@@ -41,7 +41,7 @@ EXAMPLES:
 
 | Year | Features | Bugfixes | Changes | Maintenance | Docs | Total |
 |------|----------|----------|---------|-------------|------|-------|
-| 2026 | 80       | 85       | 30 | 81 | 40 | 317   |
+| 2026 | 80       | 86       | 30 | 81 | 40 | 318   |
 | 2025 | 60       | 68       | 22 | 71 | 36 | 256   |
 | 2024 | 12       | 17       | 1 | 12 | 1 | 43    |
 | 2023 | 11       | 14       | 3 | 9 | 1 | 38    |
@@ -53,6 +53,11 @@ EXAMPLES:
 | 2017 | 9        | 0        | 3 | 2 | 0 | 14    |
 
 ## 2026
+
+### 17 Jul 2026
+
+- [bugfix] Corrected the building material properties used by DyOHM
+  - Storage-heat methods 6 and 8 now calculate building DyOHM coefficients from material layer 0 of `land_cover.bldgs` instead of a SPARTACUS wall.
 
 ### 13 Jul 2026